testing AlarmManager of ScheduledRecordingService
