--- app/src/androidTest/java/com/danielkim/soundrecorder/AlarmManagerTest.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2017. Software developed by iClaude. All rights reserved.
- */
-
-/*
-package com.danielkim.soundrecorder;
-
-import android.app.AlarmManager;
-import android.content.Context;
-
-import org.junit.Before;
-import org.junit.runner.RunWith;
-import org.robolectric.RobolectricTestRunner;
-import org.robolectric.RuntimeEnvironment;
-import org.robolectric.annotation.Config;
-import org.robolectric.shadows.ShadowAlarmManager;
-
-import static org.robolectric.Shadows.shadowOf;
-
-*/
-/**
- * Created by Claudio on 12/07/2017.
- *//*
-
-
-@RunWith(RobolectricTestRunner.class)
-@Config(constants = BuildConfig.class, sdk = 21)
-public class AlarmManagerTest {
-
-    private Context context;
-    private ShadowAlarmManager shadowAlarmManager;
-    private AlarmManager alarmManager;
-
-    @Before
-    public void setUp() throws Exception {
-        context = RuntimeEnvironment.application.getApplicationContext();
-        AlarmManager alarmManager = (AlarmManager) RuntimeEnvironment.application.getSystemService(Context.ALARM_SERVICE);
-        shadowAlarmManager = shadowOf(alarmManager);
-    }
-}
-*/
